--- a/.travis.yml
+++ b/.travis.yml
@@ -1,6 +1,6 @@
 language: node_js
 node_js:
-  - "6"
-  - "8"
+  - "14"
+  - "16"
   - "stable"
 sudo: false