--- a/appveyor.yml
+++ b/appveyor.yml
@@ -7,8 +7,8 @@ init:
 # Test against these versions of Node.js.
 environment:
   matrix:
-    - nodejs_version: "6"
-    - nodejs_version: "8"
+    - nodejs_version: "14"
+    - nodejs_version: "16"
 
 # Install scripts. (runs after repo cloning)
 install: