--- a/package.json
+++ b/package.json
@@ -18,7 +18,7 @@
     "entry.js"
   ],
   "dependencies": {
-    "fs-updater": "^1.0.4",
+    "fs-updater": "RuslanZavacky/node-fs-updater#optimizations",
     "heimdalljs": "^0.2.5"
   },
   "devDependencies": {