--- a/package.json
+++ b/package.json
@@ -27,7 +27,7 @@
     "fs-extra": "^5.0.0",
     "mocha": "^5.0.1",
     "mocha-eslint": "^4.1.0",
-    "rimraf": "^2.6.2"
+    "rimraf": "^3.0.2"
   },
   "scripts": {
     "test": "mocha",