This is an additive PR for the generic effort that I am trying to pursue to speed up merge-trees for the bigger projects.

Reference: https://github.com/broccolijs/node-fs-updater/pull/4

Here I've fixed a typo in the benchmark and increased it's numbers a bit, to get bigger set of data emulating more "work".

Another bigger question that I had is to figure out a way to go from recursion to iteration and potentially save a bunch of memory used while doing iterations. It does seem like that require much more comprehensive refactoring, which deem quite challenging, but I might try that as well.