@RuslanZavacky 2022-02-01
Here are some interesting numbers that I've got interim:

MacBook Pro (16-inch, 2019)
Processor: 2.4 GHz 8-Core Intel Core i9
Memory: 32 GB 2667 MHz DDR4
Graphics: AMD Radeon Pro 5300M 4 GB
Intel UHD Graphics 630 1536 MB

Running updated benchmark:

| type                                | Intel (this branch) | Intel (master) | M1 (this branch) | M1 (master) |
|-------------------------------------|---------------------|----------------|------------------|-------------|
| initial merge, canSymlink=false     | 19.533s             | 17.364s        | 3.935s           | 4.159s      |
| incremental merge, canSymlink=false | 3.425s              | 3.153s         | 1.006s           | 994.372ms   |
| initial merge, canSymlink=null      | 9.253s              | 8.447s         | 2.702s           | 2.689s      |
| incremental merge, canSymlink=null  | 485.557ms           | 647.591ms      | 127.09ms         | 253.791ms   |